Dropping mathematical approach. Doing it dumb

+++ 2019/cpp/Day 3/include/solution.hpp
@@ -77,7 +77,7 @@ class WireSet {
     public:
         WireSet(std::vector<Wire> wires);
 
-        double getManhatanDistanceToClosestCross();
+        long getManhatanDistanceToClosestCross();
 
     private:
         std::unordered_map<Point, int> wireMap;
